+++ modules/config/default.js
@@ -69,6 +69,5 @@ export const settings = {
   valueSourcesPopupTitle: "Select value source",
   removeRuleConfirmOptions: null,
   removeGroupConfirmOptions: null,
-  useGroupsAsArrays: true,
   defaultGroupConjunction: "AND"
 };